Better FileUtil#findClass
Issue #727 's CME roots come from FileUtil. This can be due to the previous
implementation having 2 lists linking each other. This is removing 1 of the
lists and handling on-the-go while also cleaning up the code a bit.

Oh, and for the CloudExpansionManager change: I put it on the wrong
spot the first time :P so yea